% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{analysis_emax}
\alias{analysis_emax}
\title{analysis_emax}
\usage{
analysis_emax(design, doses, schedules, n_pat, likelihood, threshold)
}
\arguments{
\item{design}{teh design matrix to be evaluated}

\item{doses}{dose levels}

\item{schedules}{schedule levels}

\item{n_pat}{sample size to be evaluated}

\item{likelihood}{either "linear" or "emax" - true model}

\item{threshold}{placebo adjusted threshold, teh effect has to pass to being considered an MED}
}
\value{
response_design containing MED, effect, AIC, BIC as attributes
}
\description{
analysis_emax
}
