% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{analysis_linear}
\alias{analysis_linear}
\title{analysis_linear}
\usage{
analysis_linear(design, doses, schedules, n_pat, likelihood, threshold)
}
\arguments{
\item{design}{the design matrix to be evaluated}

\item{doses}{the doses}

\item{schedules}{the schedules}

\item{n_pat}{the overall to be evaluated}

\item{likelihood}{either "linear" or "emax" - true model}

\item{threshold}{placebo adjusted threshold, the effect has to pass to being considered an MED}
}
\value{
response_design containing MED, effect, AIC, BIC as attributes
}
\description{
analysis_linear
}
