% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_MSE_RMSE_across_scenarios}
\alias{calc_MSE_RMSE_across_scenarios}
\title{Title}
\usage{
calc_MSE_RMSE_across_scenarios(results, scenario, interim)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{interim}{tbd}
}
\value{
tbd
}
\description{
Title
}
