% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{calculate_columns}
\alias{calculate_columns}
\title{calculate_columns}
\usage{
calculate_columns(scenario_row, trialsims)
}
\arguments{
\item{scenario_row}{the row of the scenario table}

\item{trialsims}{the result dataset - potentially a targets object}
}
\value{
result dataframe containing all relevent information for that scenario
}
\description{
calculate_columns
}
