% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calculate_MED_emax}
\alias{calculate_MED_emax}
\title{Title}
\usage{
calculate_MED_emax(
  E0,
  alpha1,
  alpha2,
  beta,
  delta1,
  delta2,
  efficacy_threshold,
  doses,
  schedules,
  med,
  obs_effect
)
}
\arguments{
\item{E0}{tbd}

\item{alpha1}{tbd}

\item{alpha2}{tbd}

\item{beta}{tbd}

\item{delta1}{tbd}

\item{delta2}{tbd}

\item{efficacy_threshold}{tbd}

\item{doses}{tbd}

\item{schedules}{tbd}

\item{med}{tbd}

\item{obs_effect}{tbd}
}
\value{
tbd
}
\description{
Title
}
