% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calculate_MED_linear}
\alias{calculate_MED_linear}
\title{Title}
\usage{
calculate_MED_linear(
  beta0,
  beta1,
  beta2,
  beta3,
  efficacy_threshold,
  doses,
  schedules,
  med,
  obs_effect
)
}
\arguments{
\item{beta0}{tbd}

\item{beta1}{tbd}

\item{beta2}{tbd}

\item{beta3}{tbd}

\item{efficacy_threshold}{tbd}

\item{doses}{tbd}

\item{schedules}{tbd}

\item{med}{tbd}

\item{obs_effect}{tbd}
}
\value{
tbd
}
\description{
Title
}
