% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{gen_likelihood_data_emax}
\alias{gen_likelihood_data_emax}
\title{gen_likelihood_data_emax}
\usage{
gen_likelihood_data_emax(
  data_mat,
  E0,
  alpha1,
  alpha2,
  delta1,
  delta2,
  beta,
  sig_mu,
  sig_sd,
  n_pat
)
}
\arguments{
\item{data_mat}{matrix}

\item{initial_E0}{prior of the E0}

\item{initial_alpha1}{prior of the doses EMax}

\item{initial_alpha2}{prior of the schedules EMax}

\item{initial_delta1}{prior of the doses ED50}

\item{initial_delta2}{prior of the schedules ED50}

\item{initial_beta}{interaction}

\item{initial_sigma_mu}{mean of epsilon}

\item{initial_sigma_sd}{sigma}
}
\value{
res a list containing patient level data
}
\description{
gen_likelihood_data_emax
}
