% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{gen_likelihood_data_linear}
\alias{gen_likelihood_data_linear}
\title{Title}
\usage{
gen_likelihood_data_linear(
  data_mat,
  beta0,
  beta1,
  beta2,
  beta3,
  sig_mu,
  sig_sd,
  n_pat
)
}
\arguments{
\item{data_mat}{tbd}

\item{initial_beta0}{tbd}

\item{initial_beta1}{tbd}

\item{initial_beta2}{tbd}

\item{initial_beta3}{tbd}

\item{initial_sigma_mu}{tbd}

\item{initial_sigma_sd}{tbd}
}
\value{
tbd
}
\description{
Title
}
