% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_custom}
\alias{generate_custom}
\title{generate_custom}
\usage{
generate_custom(doses, schedules, n_pat)
}
\arguments{
\item{doses}{dose levels}

\item{schedules}{schedule levels}

\item{n_pat}{sample size to be allocated across the factorial space}
}
\value{
matrix
}
\description{
generate_custom
}
