% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_custom_design}
\alias{generate_custom_design}
\title{generate_custom_design}
\usage{
generate_custom_design(n_pat, post_probs)
}
\arguments{
\item{n_pat}{number of patients allocated across the factorial space}

\item{post_probs}{posterior probabilities of the grid the patients should be allocated on}
}
\value{
matrix
}
\description{
generate_custom_design
}
