% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_full_factorial_design}
\alias{generate_full_factorial_design}
\title{generate_full_factorial_design}
\usage{
generate_full_factorial_design(doses, schedules, n_pat)
}
\arguments{
\item{doses}{the dose levels}

\item{schedules}{the schedules}

\item{n_pat}{sample size to be allocated across the factorial space}
}
\value{
matrix
}
\description{
generate_full_factorial_design
}
