% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_optimal_design_emax}
\alias{generate_optimal_design_emax}
\title{Title}
\usage{
generate_optimal_design_emax(
  doses,
  schedules,
  n_pat,
  criterion = "D",
  random_order = TRUE,
  emax_d,
  ed50_d,
  emax_s,
  ed50_s,
  interaction = TRUE
)
}
\arguments{
\item{doses}{tbd}

\item{schedules}{tbd}

\item{n_pat}{tbd}

\item{criterion}{tbd}

\item{random_order}{tbd}

\item{emax_d}{tbd}

\item{ed50_d}{tbd}

\item{emax_s}{tbd}

\item{ed50_s}{tbd}

\item{interaction}{tbd}
}
\value{
matrix
}
\description{
Title
}
