% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_optimal_design_linear}
\alias{generate_optimal_design_linear}
\title{Title}
\usage{
generate_optimal_design_linear(
  doses,
  schedules,
  n_pat,
  criterion = "D",
  random_order = TRUE,
  interaction = TRUE
)
}
\arguments{
\item{doses}{tbd}

\item{schedules}{tbd}

\item{n_pat}{tbd}

\item{criterion}{tbd}

\item{random_order}{tbd}
}
\value{
matrix
}
\description{
Title
}
