% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_ABIC_score}
\alias{get_ABIC_score}
\title{get_ABIC_score}
\usage{
get_ABIC_score(results, scenario, abic, med, method, GoNoGo_Threshold)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{abic}{location of AIC/BIC values}

\item{med}{location of MED}

\item{method}{for which option sum should get calulated. Options are:
"gonogo", "corr_med", "mse", "rmse", "pat_alloc"}

\item{GoNoGo_Threshold}{gonogo threshold to be passed, e.g., 0.5}
}
\value{
sum
}
\description{
get_ABIC_score
}
