% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_AIC_BIC_scores_across}
\alias{get_AIC_BIC_scores_across}
\title{get_AIC_BIC_scores_across}
\usage{
get_AIC_BIC_scores_across(results, scenario, interim, gonogo_threshold)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{interim}{boolean, indicating whether an interim analysis has been conducted}

\item{gonogo_threshold}{gonogo threshold to be passed, e.g., 0.5}
}
\value{
dataframe
}
\description{
get_AIC_BIC_scores_across
}
