% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_all_correct_med_est}
\alias{get_all_correct_med_est}
\title{get_all_correct_med_est}
\usage{
get_all_correct_med_est(results, scenario, interim)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{interim}{boolean, indicating whether an interim analysis has been conducted}
}
\value{
list
}
\description{
get_all_correct_med_est
}
