% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_alloc}
\alias{get_alloc}
\title{get_alloc}
\usage{
get_alloc(results, scenario, mat)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{mat}{matrix to be looked at}
}
\value{
sum
}
\description{
get_alloc
}
