% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_bayesfactor}
\alias{get_bayesfactor}
\title{Title}
\usage{
get_bayesfactor(all_mcmc, w_prior = c(1/2, 1/2), data)
}
\arguments{
\item{all_mcmc}{tbd}

\item{w_prior}{tbd}

\item{data}{tbd}
}
\value{
list
}
\description{
Title
}
