% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_effect}
\alias{get_effect}
\title{Title}
\usage{
get_effect(results, scenario, likelihood, var)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{likelihood}{tbd}

\item{var}{tbd}
}
\value{
tbd
}
\description{
Title
}
