% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_effect_true}
\alias{get_effect_true}
\title{Title}
\usage{
get_effect_true(results, scenario, var)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{var}{tbd}
}
\value{
tbd
}
\description{
Title
}
