% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_GoNoGo}
\alias{get_GoNoGo}
\title{get_GoNoGo}
\usage{
get_GoNoGo(results, scenario, GoNoGo_Threshold, interim)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{GoNoGo_Threshold}{gonogo threshold to be passed, e.g., 0.5}

\item{interim}{boolean, indicating whether an interim analysis has been conducted}
}
\value{
list
}
\description{
get_GoNoGo
}
