% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_MED_alloc_no_interim}
\alias{get_MED_alloc_no_interim}
\title{get_MED_alloc_no_interim}
\usage{
get_MED_alloc_no_interim(results, scenario, mat)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{mat}{matrix to be looked at}
}
\value{
sum
}
\description{
get_MED_alloc_no_interim
}
