% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_MED_CRE}
\alias{get_MED_CRE}
\title{get_MED_CRE}
\usage{
get_MED_CRE(doses, schedules, lb, threshold)
}
\arguments{
\item{doses}{the doses}

\item{schedules}{the schedules}

\item{lb}{lower boundary of the posterior samples}

\item{threshold}{placebo adjusted threshold, teh effect has to pass to being considered an MED}
}
\value{
list
}
\description{
get_MED_CRE
}
