% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_post_for_quantile_samples_emax}
\alias{get_post_for_quantile_samples_emax}
\title{get_post_for_quantile_samples_emax}
\usage{
get_post_for_quantile_samples_emax(samples, quantile = 0.5)
}
\arguments{
\item{samples}{posterior samples to be evaluated}

\item{quantile}{the quantile to be used}
}
\value{
quantiles  respective chosen posterior samples
}
\description{
get_post_for_quantile_samples_emax
}
