% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_post_for_quantile_samples_linear}
\alias{get_post_for_quantile_samples_linear}
\title{get_post_for_quantile_samples_linear}
\usage{
get_post_for_quantile_samples_linear(samples, quantile = 0.5)
}
\arguments{
\item{samples}{posterior samples}

\item{quantile}{quantile to be evaluated}
}
\value{
quantiles  respective posterior quantiles
}
\description{
get_post_for_quantile_samples_linear
}
