% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_resp}
\alias{get_resp}
\title{get_resp}
\usage{
get_resp(doses, schedules, quantiles, emax = FALSE, n_pat, patient_data)
}
\arguments{
\item{doses}{tbd}

\item{schedules}{tbd}

\item{quantiles}{tbd}

\item{emax}{tbd}
}
\value{
resp
}
\description{
get_resp
}
