% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenarios.R
\name{get_scenario}
\alias{get_scenario}
\title{get_scenario}
\usage{
get_scenario(scenario_label)
}
\arguments{
\item{scenario_label}{returns the chosen scenario row}
}
\value{
df
}
\description{
get_scenario
}
