% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{linear}
\alias{linear}
\title{Title}
\usage{
linear(
  data = n_matrix,
  doses,
  schedules,
  n_chain = 3,
  n_burn = 1000,
  n_sample = 1000,
  n_thin = 5,
  n_pat,
  efficacy_threshold,
  likelihood_data
)
}
\arguments{
\item{data}{tbd}

\item{n_chain}{tbd}

\item{n_burn}{tbd}

\item{n_sample}{tbd}

\item{n_thin}{tbd}
}
\value{
tbd
}
\description{
Title
}
