% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_final_trial}
\alias{simulate_final_trial}
\title{simulate_final_trial}
\usage{
simulate_final_trial(
  scenario_label = "null",
  doses = c(0, 4, 5, 6, 9),
  schedules = c(1, 3, 4),
  threshold = 8,
  alloc_mat = 1,
  n_pat = 200,
  interim = TRUE,
  likelihood = "linear",
  E0 = -12.8,
  alpha1 = 10,
  alpha2 = 10,
  delta1 = 3,
  delta2 = 1,
  beta = 0,
  sig_mu = 1,
  sig_sd = 1,
  beta0 = -12.8,
  beta1 = 0.1319648,
  beta2 = 0.1319648,
  beta3 = 0
)
}
\arguments{
\item{scenario_label}{the label}

\item{doses}{tbd}

\item{schedules}{tbd}

\item{threshold}{tbd}

\item{alloc_mat}{tbd}

\item{n_pat}{tbd}

\item{interim}{tbd}

\item{likelihood}{tbd}

\item{E0}{tbd}

\item{alpha1}{tbd}

\item{alpha2}{tbd}

\item{delta1}{tbd}

\item{delta2}{tbd}

\item{beta}{tbd}

\item{sig_mu}{tbd}

\item{sig_sd}{tbd}

\item{beta0}{tbd}

\item{beta1}{tbd}

\item{beta2}{tbd}

\item{beta3}{tbd}
}
\value{
df
}
\description{
simulate_final_trial
}
