% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{analysis_emax}
\alias{analysis_emax}
\title{Title}
\usage{
analysis_emax(design, doses, schedules, n_pat, likelihood, threshold)
}
\arguments{
\item{design}{tbd}

\item{doses}{tbd}

\item{schedules}{tbd}

\item{n_pat}{tbd}

\item{likelihood}{tbd}

\item{threshold}{tbd}
}
\value{
design
}
\description{
Title
}
