% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_MSE_RMSE_across_scenarios}
\alias{calc_MSE_RMSE_across_scenarios}
\title{calc_MSE_RMSE_across_scenarios}
\usage{
calc_MSE_RMSE_across_scenarios(results, scenario, interim)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{interim}{boolean, whether interim was conducted or not}
}
\value{
list containg MSE and RMSE
}
\description{
calc_MSE_RMSE_across_scenarios
}
