% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{calculate_columns}
\alias{calculate_columns}
\title{Title}
\usage{
calculate_columns(scenario_row, trialsims)
}
\arguments{
\item{scenario_row}{tbd}

\item{trialsims}{tbd}
}
\value{
tbd
}
\description{
Title
}
