% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{calculate_MED_emax}
\alias{calculate_MED_emax}
\title{calculate_MED_emax}
\usage{
calculate_MED_emax(
  E0,
  alpha1,
  alpha2,
  beta,
  delta1,
  delta2,
  efficacy_threshold,
  doses,
  schedules,
  med,
  obs_effect
)
}
\arguments{
\item{E0}{E0}

\item{alpha1}{EMax of doses}

\item{alpha2}{EMax of schedules}

\item{beta}{interaction}

\item{delta1}{ED50 of doses}

\item{delta2}{ED50 of schedules}

\item{efficacy_threshold}{placebo adjusted threshold to be passed to be considered an MED}

\item{doses}{doses}

\item{schedules}{schedules}

\item{med}{an empty matrix the med gets allocated on}

\item{obs_effect}{an empty matrix the observed effects gets allocated on}
}
\value{
med and effect as attribute
}
\description{
calculate_MED_emax
}
