% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{emax}
\alias{emax}
\title{emax}
\usage{
emax(
  data = n_matrix,
  doses,
  schedules,
  n_chain = 3,
  n_burn = 1000,
  n_sample = 1000,
  n_thin = 5,
  n_pat,
  efficacy_threshold,
  likelihood_data
)
}
\arguments{
\item{data}{matrix}

\item{doses}{doses}

\item{schedules}{schedules}

\item{n_chain}{number of chains to be used for the JAGS model, default 3}

\item{n_burn}{number of burn samples to be used for the JAGS model, default 1000}

\item{n_sample}{number of samples to be used for the JAGS model, default 1000}

\item{n_thin}{thinning rate to be used for the JAGS model, default 5}

\item{n_pat}{sample size}

\item{efficacy_threshold}{placebo adjusted threshold to be passed to be considered an MED}

\item{likelihood_data}{likelihood patient level data}
}
\value{
res, containing the patient level responses, as attributes also the MED, effect, all samples and patient level data
}
\description{
emax
}
