% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{gen_likelihood_data_emax}
\alias{gen_likelihood_data_emax}
\title{Title}
\usage{
gen_likelihood_data_emax(
  data_mat,
  E0,
  alpha1,
  alpha2,
  delta1,
  delta2,
  beta,
  sig_mu,
  sig_sd,
  n_pat
)
}
\arguments{
\item{data_mat}{tbd}

\item{initial_E0}{tbd}

\item{initial_alpha1}{tbd}

\item{initial_alpha2}{tbd}

\item{initial_delta1}{tbd}

\item{initial_delta2}{tbd}

\item{initial_beta}{tbd}

\item{initial_sigma_mu}{tbd}

\item{initial_sigma_sd}{tbd}
}
\value{
tbd
}
\description{
Title
}
