% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{gen_likelihood_data_linear}
\alias{gen_likelihood_data_linear}
\title{gen_likelihood_data_linear}
\usage{
gen_likelihood_data_linear(
  data_mat,
  beta0,
  beta1,
  beta2,
  beta3,
  sig_mu,
  sig_sd,
  n_pat
)
}
\arguments{
\item{data_mat}{matrix}

\item{initial_beta0}{prior of beta0}

\item{initial_beta1}{prior of coefficient for doses}

\item{initial_beta2}{prior of coefficient for schedukles}

\item{initial_beta3}{prior for interaction}

\item{initial_sigma_mu}{prior for mean of epsilon}

\item{initial_sigma_sd}{prior of sigma}
}
\value{
res a list containing patient level data
}
\description{
gen_likelihood_data_linear
}
