% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_custom_design}
\alias{generate_custom_design}
\title{Title}
\usage{
generate_custom_design(n_pat, post_probs)
}
\arguments{
\item{n_pat}{tbd}

\item{post_probs}{tbd}
}
\value{
matrix
}
\description{
Title
}
