% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_full_factorial_design}
\alias{generate_full_factorial_design}
\title{Title}
\usage{
generate_full_factorial_design(doses, schedules, n_pat)
}
\arguments{
\item{doses}{tbd}

\item{schedules}{tbd}

\item{n_pat}{tbd}
}
\value{
matrix
}
\description{
Title
}
