% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_optimal_design_linear}
\alias{generate_optimal_design_linear}
\title{generate_optimal_design_linear}
\usage{
generate_optimal_design_linear(
  doses,
  schedules,
  n_pat,
  criterion = "D",
  random_order = TRUE,
  interaction = TRUE
)
}
\arguments{
\item{doses}{doses}

\item{schedules}{schedules}

\item{n_pat}{sample size to be allocated across the factorial space}

\item{criterion}{optimality criteria, default "D"}

\item{random_order}{boolean, default TRUE}

\item{interaction}{boolean, default TRUE}
}
\value{
matrix
}
\description{
generate_optimal_design_linear
}
