% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_ABIC_score}
\alias{get_ABIC_score}
\title{Title}
\usage{
get_ABIC_score(results, scenario, abic, med, method, GoNoGo_Threshold)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{abic}{tbd}

\item{med}{tbd}

\item{method}{tbd}

\item{GoNoGo_Threshold}{tbd}
}
\value{
tbd
}
\description{
Title
}
