% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_AIC_BIC_scores_across}
\alias{get_AIC_BIC_scores_across}
\title{Title}
\usage{
get_AIC_BIC_scores_across(results, scenario, interim, gonogo_threshold)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{interim}{tbd}

\item{gonogo_threshold}{tbd}
}
\value{
tbd
}
\description{
Title
}
