% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_all_correct_med_est}
\alias{get_all_correct_med_est}
\title{Title}
\usage{
get_all_correct_med_est(results, scenario, interim)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{interim}{tbd}
}
\value{
tbd
}
\description{
Title
}
