% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_correct_med_est}
\alias{get_correct_med_est}
\title{get_correct_med_est}
\usage{
get_correct_med_est(results, scenario, likelihood, var)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{likelihood}{tbd}

\item{var}{tbd}
}
\value{
tbd
}
\description{
get_correct_med_est
}
