% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_effect}
\alias{get_effect}
\title{get_effect}
\usage{
get_effect(results, scenario, likelihood, var)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{likelihood}{true model, either "linear" or "emax"}

\item{var}{numeric value of the value to be looked at}
}
\value{
sum
}
\description{
get_effect
}
