% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_effect_true}
\alias{get_effect_true}
\title{get_effect_true}
\usage{
get_effect_true(results, scenario, var)
}
\arguments{
\item{results}{results dataset}

\item{scenario}{numeric value of scenario}

\item{var}{variable to be looked at, numeric}
}
\value{
sum numeric value of the effect
}
\description{
get_effect_true
}
