% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_GoNoGo}
\alias{get_GoNoGo}
\title{Title}
\usage{
get_GoNoGo(results, scenario, GoNoGo_Threshold, interim)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{GoNoGo_Threshold}{tbd}

\item{interim}{tbd}
}
\value{
tbd
}
\description{
Title
}
