% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_MED_alloc_no_interim}
\alias{get_MED_alloc_no_interim}
\title{Title}
\usage{
get_MED_alloc_no_interim(results, scenario, mat)
}
\arguments{
\item{results}{tbd}

\item{scenario}{tbd}

\item{mat}{tbd}
}
\value{
tbd
}
\description{
Title
}
