% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_MED_CRE}
\alias{get_MED_CRE}
\title{Title}
\usage{
get_MED_CRE(doses, schedules, lb, threshold)
}
\arguments{
\item{doses}{tbd}

\item{schedules}{tbd}

\item{lb}{tbd}

\item{threshold}{tbd}
}
\value{
list
}
\description{
Title
}
