% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_post_for_quantile_samples_linear}
\alias{get_post_for_quantile_samples_linear}
\title{Title}
\usage{
get_post_for_quantile_samples_linear(samples, quantile = 0.5)
}
\arguments{
\item{samples}{tbd}

\item{quantile}{tbd}
}
\value{
quantiles  tbd
}
\description{
Title
}
