% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_resp}
\alias{get_resp}
\title{get_resp}
\usage{
get_resp(doses, schedules, quantiles, emax = FALSE, n_pat, patient_data)
}
\arguments{
\item{doses}{the dose levels to be evaluated}

\item{schedules}{the schedules to be evaluated}

\item{quantiles}{the posterior quantiles to be passed}

\item{emax}{boolean, whether this is an emax model or not}
}
\value{
resp containing the responses on patient level and the AIC/BIC as attributes
}
\description{
get_resp
}
