% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_final_trial}
\alias{simulate_final_trial}
\title{simulate_final_trial}
\usage{
simulate_final_trial(
  scenario_label = "null",
  doses = c(0, 4, 5, 6, 9),
  schedules = c(1, 3, 4),
  threshold = 8,
  alloc_mat = 1,
  n_pat = 200,
  interim = TRUE,
  likelihood = "linear",
  E0 = -12.8,
  alpha1 = 10,
  alpha2 = 10,
  delta1 = 3,
  delta2 = 1,
  beta = 0,
  sig_mu = 1,
  sig_sd = 1,
  beta0 = -12.8,
  beta1 = 0.1319648,
  beta2 = 0.1319648,
  beta3 = 0
)
}
\arguments{
\item{scenario_label}{the label fo the scenario}

\item{doses}{the dose levels to be investigated}

\item{schedules}{the schedules do be used}

\item{threshold}{the threshold the placebo adjusted effect has to pass in order to become an MED}

\item{alloc_mat}{the allocation matrix to be used}

\item{n_pat}{the overall sample size of the trial}

\item{interim}{boolean, whether an interim analysis should be performed or not}

\item{likelihood}{character string - either "linear" or "emax" - represents the true underlying model}

\item{E0}{the $E_0$ value of the EMAX model}

\item{alpha1}{$E_{max}$ value of the doses}

\item{alpha2}{$E_{max}$ value of the schedules}

\item{delta1}{$ED_{50}$ value of the doses}

\item{delta2}{$ED_{50}$ value of the schedules}

\item{beta}{interaction term for the EMAX model}

\item{sig_mu}{mean of the error term}

\item{sig_sd}{sigma of the error term}

\item{beta0}{intercept of the linear model}

\item{beta1}{coefficient of the doses}

\item{beta2}{coefficient of the schedules}

\item{beta3}{interaction term of the linear model}
}
\value{
out a tibble containing the results
}
\description{
simulate_final_trial
}
